unify the code and file names for all the ansible creator functionalities
