+++ .vscodeignore
@@ -15,12 +15,13 @@
 !media/codicons/codicon.css
 !media/codicons/codicon.ttf
 !media/codicons/codicon.css
+!media/baseStyles/baseFormStyle.css
 !media/contentCreator/icons/ansible-creator-create.png
 !media/contentCreator/icons/ansible-creator-init.png
 !media/contentCreator/icons/ansible-creator-sample.png
 !media/contentCreator/icons/ansible-logo-red.png
-!media/contentCreator/scaffoldCollectionPageStyle.css
-!media/contentCreator/scaffoldAnsibleProjectPageStyle.css
+!media/contentCreator/createAnsibleCollectionPageStyle.css
+!media/contentCreator/createAnsibleProjectPageStyle.css
 !media/contentCreator/welcomePageStyle.css
 !media/lightspeedExplorerView/style.css
 !media/playbookGeneration/playbookGeneration.css
@@ -31,7 +32,8 @@
 !out/client/webview/apps/lightspeed/playbookExplanation/main.js
 !out/client/webview/apps/lightspeed/playbookGeneration/main.js
 !out/client/webview/apps/contentCreator/welcomePageApp.js
-!out/client/webview/apps/contentCreator/scaffoldCollectionPageApp.js
+!out/client/webview/apps/contentCreator/createAnsibleCollectionPageApp.js
+!out/client/webview/apps/contentCreator/createAnsibleProjectPageApp.js
 !out/server/src/server.js
 !package.json
 !package.nls.json
