+++ media/contentCreator/createAnsibleProjectPageStyle.css
@@ -1,21 +1,4 @@
-body {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-section {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 20px;
-    /* width: 600px; */
-    /* width: 70%; */
-    max-width: 600px;
-    /* margin: auto; */
-}
+@import url(../baseStyles/baseFormStyle.css);
 
 .container {
     display: flex;
@@ -36,6 +19,18 @@ vscode-text-area {
     margin-bottom: 6px;
 }
 
+.scm-org-project-div {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+}
+
+#scm-org-name, #scm-project-name {
+    width:49%;
+    display:inline-block;
+}
+
 .checkbox-div {
     display: flex; /* Use flexbox */
     flex-direction: column; /* Arrange child elements vertically */
@@ -49,12 +44,17 @@ vscode-text-area {
     flex-direction: row; /* Arrange child elements vertically */
     margin-top: 12px;
     margin-bottom: 30px;
+    width: 100%;
 }
 
-.full-collection-name {
+vscode-dropdown {
+    width: 200px;
+}
+
+.full-collection-path {
     display: flex; /* Use flexbox */
     flex-direction: row; /* Arrange child elements vertically */
-    color: grey;
+    color: var(--vscode-descriptionForeground);
 }
 
 .group-buttons {
@@ -71,7 +71,7 @@ vscode-button {
 }
 
 vscode-checkbox i {
-    color: gray;
+    color: var(--vscode-descriptionForeground);
     font-size: small;
 }
 
@@ -101,7 +101,8 @@ vscode-checkbox i {
     display: none;
     flex-direction: column;
     border-style: dotted;
-    border-color: #444444;;
+    border-color: var(--focus-border);;
+    border-width: 0.5px;
     padding: 8px;
 }
