+++ package.json
@@ -214,12 +214,12 @@
         "title": "Install Ansible Content Creator"
       },
       {
-        "command": "ansible.content-creator.scaffold-ansible-collection",
-        "title": "Scaffold Ansible Collection"
+        "command": "ansible.content-creator.create-ansible-collection",
+        "title": "Create Ansible Collection"
       },
       {
-        "command": "ansible.content-creator.scaffold-ansible-project",
-        "title": "Scaffold Ansible Project"
+        "command": "ansible.content-creator.create-ansible-project",
+        "title": "Create Ansible Project"
       },
       {
         "command": "ansible.content-creator.create",
