+++ webpack.config.ts
@@ -98,20 +98,6 @@ const contentCreatorMenuWebviewConfig = {
   },
 };
 
-const contentCreatorInitWebviewConfig = {
-  ...config,
-  target: ["web", "es2020"],
-  entry: "./src/webview/apps/contentCreator/scaffoldCollectionPageApp.ts",
-  experiments: { outputModule: true },
-  output: {
-    path: path.resolve(__dirname, "out"),
-    filename:
-      "./client/webview/apps/contentCreator/scaffoldCollectionPageApp.js",
-    libraryTarget: "module",
-    chunkFormat: "module",
-  },
-};
-
 const playbookExplorerWebviewConfig = {
   ...config,
   target: ["web", "es2020"],
@@ -151,15 +137,29 @@ const playbookExplanationWebviewConfig = {
   },
 };
 
-const scaffoldAnsibleProjectWebviewConfig = {
+const createAnsibleCollectionWebviewConfig = {
   ...config,
   target: ["web", "es2020"],
-  entry: "./src/webview/apps/contentCreator/scaffoldAnsibleProjectPageApp.ts",
+  entry: "./src/webview/apps/contentCreator/createAnsibleCollectionPageApp.ts",
   experiments: { outputModule: true },
   output: {
     path: path.resolve(__dirname, "out"),
     filename:
-      "./client/webview/apps/contentCreator/scaffoldAnsibleProjectPageApp.js",
+      "./client/webview/apps/contentCreator/createAnsibleCollectionPageApp.js",
+    libraryTarget: "module",
+    chunkFormat: "module",
+  },
+};
+
+const createAnsibleProjectWebviewConfig = {
+  ...config,
+  target: ["web", "es2020"],
+  entry: "./src/webview/apps/contentCreator/createAnsibleProjectPageApp.ts",
+  experiments: { outputModule: true },
+  output: {
+    path: path.resolve(__dirname, "out"),
+    filename:
+      "./client/webview/apps/contentCreator/createAnsibleProjectPageApp.js",
     libraryTarget: "module",
     chunkFormat: "module",
   },
@@ -169,9 +169,9 @@ module.exports = [
   config,
   webviewConfig,
   contentCreatorMenuWebviewConfig,
-  contentCreatorInitWebviewConfig,
+  createAnsibleCollectionWebviewConfig,
   playbookExplorerWebviewConfig,
   playbookGenerationWebviewConfig,
   playbookExplanationWebviewConfig,
-  scaffoldAnsibleProjectWebviewConfig,
+  createAnsibleProjectWebviewConfig,
 ];
